+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Personalized, AI assistant that runs with your engineering teams"
 authors = ["Perfeed Core"]
 license = "MIT"
 readme = "README.md"
-packages = [{include = "src"}]
+packages = [{include = "perfeed"}]
 
 [tool.poetry.dependencies]
 python = "^3.10"
